@@ -1,3 +1,20 @@
 module github.com/graphprotocol/proto-cosmos
 
-go 1.20
+go 1.22
+
+toolchain go1.24.1
+
+require (
+	github.com/cosmos/cosmos-proto v1.0.0-beta.5
+	github.com/gogo/protobuf v1.3.2
+	github.com/streamingfast/pbgo v0.0.6-0.20250120164644-a58d8066ab4b
+	google.golang.org/protobuf v1.35.1
+)
+
+require (
+	golang.org/x/net v0.30.0 // indirect
+	golang.org/x/sys v0.26.0 // indirect
+	golang.org/x/text v0.19.0 // indirect
+	google.golang.org/genproto v0.0.0-20220808131553-a91ffa7f803e // indirect
+	google.golang.org/grpc v1.69.2 // indirect
+)
