@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.7
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: sf/cosmos/transform/v1/transform.proto
 
 package pbtransform
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -21,22 +22,19 @@ const (
 )
 
 type CombinedFilter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	EventTypeFilters   []*EventTypeFilter   `protobuf:"bytes,1,rep,name=event_type_filters,json=eventTypeFilters,proto3" json:"event_type_filters,omitempty"`
-	EventOriginFilters []*EventOriginFilter `protobuf:"bytes,2,rep,name=event_origin_filters,json=eventOriginFilters,proto3" json:"event_origin_filters,omitempty"`
-	MessageTypeFilters []*MessageTypeFilter `protobuf:"bytes,3,rep,name=message_type_filters,json=messageTypeFilters,proto3" json:"message_type_filters,omitempty"`
+	state              protoimpl.MessageState `protogen:"open.v1"`
+	EventTypeFilters   []*EventTypeFilter     `protobuf:"bytes,1,rep,name=event_type_filters,json=eventTypeFilters,proto3" json:"event_type_filters,omitempty"`
+	EventOriginFilters []*EventOriginFilter   `protobuf:"bytes,2,rep,name=event_origin_filters,json=eventOriginFilters,proto3" json:"event_origin_filters,omitempty"`
+	MessageTypeFilters []*MessageTypeFilter   `protobuf:"bytes,3,rep,name=message_type_filters,json=messageTypeFilters,proto3" json:"message_type_filters,omitempty"`
+	unknownFields      protoimpl.UnknownFields
+	sizeCache          protoimpl.SizeCache
 }
 
 func (x *CombinedFilter) Reset() {
 	*x = CombinedFilter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CombinedFilter) String() string {
@@ -47,7 +45,7 @@ func (*CombinedFilter) ProtoMessage() {}
 
 func (x *CombinedFilter) ProtoReflect() protoreflect.Message {
 	mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -84,20 +82,17 @@ func (x *CombinedFilter) GetMessageTypeFilters() []*MessageTypeFilter {
 }
 
 type EventTypeFilter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	EventTypes    []string               `protobuf:"bytes,1,rep,name=event_types,json=eventTypes,proto3" json:"event_types,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	EventTypes []string `protobuf:"bytes,1,rep,name=event_types,json=eventTypes,proto3" json:"event_types,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *EventTypeFilter) Reset() {
 	*x = EventTypeFilter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *EventTypeFilter) String() string {
@@ -108,7 +103,7 @@ func (*EventTypeFilter) ProtoMessage() {}
 
 func (x *EventTypeFilter) ProtoReflect() protoreflect.Message {
 	mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -131,20 +126,17 @@ func (x *EventTypeFilter) GetEventTypes() []string {
 }
 
 type EventOriginFilter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	EventOrigins  []string               `protobuf:"bytes,1,rep,name=event_origins,json=eventOrigins,proto3" json:"event_origins,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	EventOrigins []string `protobuf:"bytes,1,rep,name=event_origins,json=eventOrigins,proto3" json:"event_origins,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *EventOriginFilter) Reset() {
 	*x = EventOriginFilter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *EventOriginFilter) String() string {
@@ -155,7 +147,7 @@ func (*EventOriginFilter) ProtoMessage() {}
 
 func (x *EventOriginFilter) ProtoReflect() protoreflect.Message {
 	mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -178,20 +170,17 @@ func (x *EventOriginFilter) GetEventOrigins() []string {
 }
 
 type MessageTypeFilter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	MessageTypes  []string               `protobuf:"bytes,1,rep,name=message_types,json=messageTypes,proto3" json:"message_types,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	MessageTypes []string `protobuf:"bytes,1,rep,name=message_types,json=messageTypes,proto3" json:"message_types,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MessageTypeFilter) Reset() {
 	*x = MessageTypeFilter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MessageTypeFilter) String() string {
@@ -202,7 +191,7 @@ func (*MessageTypeFilter) ProtoMessage() {}
 
 func (x *MessageTypeFilter) ProtoReflect() protoreflect.Message {
 	mi := &file_sf_cosmos_transform_v1_transform_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -226,62 +215,35 @@ func (x *MessageTypeFilter) GetMessageTypes() []string {
 
 var File_sf_cosmos_transform_v1_transform_proto protoreflect.FileDescriptor
 
-var file_sf_cosmos_transform_v1_transform_proto_rawDesc = []byte{
-	0x0a, 0x26, 0x73, 0x66, 0x2f, 0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2f, 0x74, 0x72, 0x61, 0x6e,
-	0x73, 0x66, 0x6f, 0x72, 0x6d, 0x2f, 0x76, 0x31, 0x2f, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x66, 0x6f,
-	0x72, 0x6d, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x16, 0x73, 0x66, 0x2e, 0x63, 0x6f, 0x73,
-	0x6d, 0x6f, 0x73, 0x2e, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x66, 0x6f, 0x72, 0x6d, 0x2e, 0x76, 0x31,
-	0x22, 0xa1, 0x02, 0x0a, 0x0e, 0x43, 0x6f, 0x6d, 0x62, 0x69, 0x6e, 0x65, 0x64, 0x46, 0x69, 0x6c,
-	0x74, 0x65, 0x72, 0x12, 0x55, 0x0a, 0x12, 0x65, 0x76, 0x65, 0x6e, 0x74, 0x5f, 0x74, 0x79, 0x70,
-	0x65, 0x5f, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32,
-	0x27, 0x2e, 0x73, 0x66, 0x2e, 0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2e, 0x74, 0x72, 0x61, 0x6e,
-	0x73, 0x66, 0x6f, 0x72, 0x6d, 0x2e, 0x76, 0x31, 0x2e, 0x45, 0x76, 0x65, 0x6e, 0x74, 0x54, 0x79,
-	0x70, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x52, 0x10, 0x65, 0x76, 0x65, 0x6e, 0x74, 0x54,
-	0x79, 0x70, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x5b, 0x0a, 0x14, 0x65, 0x76,
-	0x65, 0x6e, 0x74, 0x5f, 0x6f, 0x72, 0x69, 0x67, 0x69, 0x6e, 0x5f, 0x66, 0x69, 0x6c, 0x74, 0x65,
-	0x72, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x29, 0x2e, 0x73, 0x66, 0x2e, 0x63, 0x6f,
-	0x73, 0x6d, 0x6f, 0x73, 0x2e, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x66, 0x6f, 0x72, 0x6d, 0x2e, 0x76,
-	0x31, 0x2e, 0x45, 0x76, 0x65, 0x6e, 0x74, 0x4f, 0x72, 0x69, 0x67, 0x69, 0x6e, 0x46, 0x69, 0x6c,
-	0x74, 0x65, 0x72, 0x52, 0x12, 0x65, 0x76, 0x65, 0x6e, 0x74, 0x4f, 0x72, 0x69, 0x67, 0x69, 0x6e,
-	0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x5b, 0x0a, 0x14, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x5f, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x18,
-	0x03, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x29, 0x2e, 0x73, 0x66, 0x2e, 0x63, 0x6f, 0x73, 0x6d, 0x6f,
-	0x73, 0x2e, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x66, 0x6f, 0x72, 0x6d, 0x2e, 0x76, 0x31, 0x2e, 0x4d,
-	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x52, 0x12, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x46, 0x69, 0x6c,
-	0x74, 0x65, 0x72, 0x73, 0x22, 0x32, 0x0a, 0x0f, 0x45, 0x76, 0x65, 0x6e, 0x74, 0x54, 0x79, 0x70,
-	0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x12, 0x1f, 0x0a, 0x0b, 0x65, 0x76, 0x65, 0x6e, 0x74,
-	0x5f, 0x74, 0x79, 0x70, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0a, 0x65, 0x76,
-	0x65, 0x6e, 0x74, 0x54, 0x79, 0x70, 0x65, 0x73, 0x22, 0x38, 0x0a, 0x11, 0x45, 0x76, 0x65, 0x6e,
-	0x74, 0x4f, 0x72, 0x69, 0x67, 0x69, 0x6e, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x12, 0x23, 0x0a,
-	0x0d, 0x65, 0x76, 0x65, 0x6e, 0x74, 0x5f, 0x6f, 0x72, 0x69, 0x67, 0x69, 0x6e, 0x73, 0x18, 0x01,
-	0x20, 0x03, 0x28, 0x09, 0x52, 0x0c, 0x65, 0x76, 0x65, 0x6e, 0x74, 0x4f, 0x72, 0x69, 0x67, 0x69,
-	0x6e, 0x73, 0x22, 0x38, 0x0a, 0x11, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70,
-	0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x12, 0x23, 0x0a, 0x0d, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0c,
-	0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x73, 0x42, 0x4d, 0x5a, 0x4b,
-	0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x67, 0x72, 0x61, 0x70, 0x68,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2d, 0x63,
-	0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2f, 0x70, 0x62, 0x2f, 0x73, 0x66, 0x2f, 0x63, 0x6f, 0x73, 0x6d,
-	0x6f, 0x73, 0x2f, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x66, 0x6f, 0x72, 0x6d, 0x2f, 0x76, 0x31, 0x3b,
-	0x70, 0x62, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x66, 0x6f, 0x72, 0x6d, 0x62, 0x06, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x33,
-}
+const file_sf_cosmos_transform_v1_transform_proto_rawDesc = "" +
+	"\n" +
+	"&sf/cosmos/transform/v1/transform.proto\x12\x16sf.cosmos.transform.v1\"\xa1\x02\n" +
+	"\x0eCombinedFilter\x12U\n" +
+	"\x12event_type_filters\x18\x01 \x03(\v2'.sf.cosmos.transform.v1.EventTypeFilterR\x10eventTypeFilters\x12[\n" +
+	"\x14event_origin_filters\x18\x02 \x03(\v2).sf.cosmos.transform.v1.EventOriginFilterR\x12eventOriginFilters\x12[\n" +
+	"\x14message_type_filters\x18\x03 \x03(\v2).sf.cosmos.transform.v1.MessageTypeFilterR\x12messageTypeFilters\"2\n" +
+	"\x0fEventTypeFilter\x12\x1f\n" +
+	"\vevent_types\x18\x01 \x03(\tR\n" +
+	"eventTypes\"8\n" +
+	"\x11EventOriginFilter\x12#\n" +
+	"\revent_origins\x18\x01 \x03(\tR\feventOrigins\"8\n" +
+	"\x11MessageTypeFilter\x12#\n" +
+	"\rmessage_types\x18\x01 \x03(\tR\fmessageTypesBMZKgithub.com/graphprotocol/proto-cosmos/pb/sf/cosmos/transform/v1;pbtransformb\x06proto3"
 
 var (
 	file_sf_cosmos_transform_v1_transform_proto_rawDescOnce sync.Once
-	file_sf_cosmos_transform_v1_transform_proto_rawDescData = file_sf_cosmos_transform_v1_transform_proto_rawDesc
+	file_sf_cosmos_transform_v1_transform_proto_rawDescData []byte
 )
 
 func file_sf_cosmos_transform_v1_transform_proto_rawDescGZIP() []byte {
 	file_sf_cosmos_transform_v1_transform_proto_rawDescOnce.Do(func() {
-		file_sf_cosmos_transform_v1_transform_proto_rawDescData = protoimpl.X.CompressGZIP(file_sf_cosmos_transform_v1_transform_proto_rawDescData)
+		file_sf_cosmos_transform_v1_transform_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_sf_cosmos_transform_v1_transform_proto_rawDesc), len(file_sf_cosmos_transform_v1_transform_proto_rawDesc)))
 	})
 	return file_sf_cosmos_transform_v1_transform_proto_rawDescData
 }
 
 var file_sf_cosmos_transform_v1_transform_proto_msgTypes = make([]protoimpl.MessageInfo, 4)
-var file_sf_cosmos_transform_v1_transform_proto_goTypes = []interface{}{
+var file_sf_cosmos_transform_v1_transform_proto_goTypes = []any{
 	(*CombinedFilter)(nil),    // 0: sf.cosmos.transform.v1.CombinedFilter
 	(*EventTypeFilter)(nil),   // 1: sf.cosmos.transform.v1.EventTypeFilter
 	(*EventOriginFilter)(nil), // 2: sf.cosmos.transform.v1.EventOriginFilter
@@ -303,61 +265,11 @@ func file_sf_cosmos_transform_v1_transform_proto_init() {
 	if File_sf_cosmos_transform_v1_transform_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_sf_cosmos_transform_v1_transform_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CombinedFilter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_sf_cosmos_transform_v1_transform_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*EventTypeFilter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_sf_cosmos_transform_v1_transform_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*EventOriginFilter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_sf_cosmos_transform_v1_transform_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MessageTypeFilter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_sf_cosmos_transform_v1_transform_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_sf_cosmos_transform_v1_transform_proto_rawDesc), len(file_sf_cosmos_transform_v1_transform_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   4,
 			NumExtensions: 0,
@@ -368,7 +280,6 @@ func file_sf_cosmos_transform_v1_transform_proto_init() {
 		MessageInfos:      file_sf_cosmos_transform_v1_transform_proto_msgTypes,
 	}.Build()
 	File_sf_cosmos_transform_v1_transform_proto = out.File
-	file_sf_cosmos_transform_v1_transform_proto_rawDesc = nil
 	file_sf_cosmos_transform_v1_transform_proto_goTypes = nil
 	file_sf_cosmos_transform_v1_transform_proto_depIdxs = nil
 }
